refactor : domain, exception 패키지로 분리
