+++ src/main/java/lotto/domain/LottoIssuanceCount.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain;
+
+import lotto.domain.exception.ErrorMessage;
 
 import java.util.ArrayList;
 import java.util.List;
