+++ src/main/java/lotto/domain/LottoProperty.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class LottoProperty {
