+++ src/main/java/lotto/domain/LottoStore.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 //구매 금액만큼 로또를 발급하는 역할을 가진 클래스
 public class LottoStore {
