+++ src/main/java/lotto/domain/WinningResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Map;
 import java.util.Set;
