+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,5 +1,7 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoProperty;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
